C51-281: Move case duration angular alterations to class
